requirements.txt
Flask>=2.0
slack-sdk>=3.0
openai>=1.0
reportlab>=3.6
python-dotenv>=0.20
psycopg2-binary>=2.9
supabase>=1.0
